backend/requirements.txt
Django>=5.0,<6.0
djangorestframework>=3.15
djangorestframework-simplejwt>=5.3
psycopg2-binary>=2.9
django-cors-headers>=4.4
python-dotenv>=1.0
celery>=5.4
redis>=5.0
drf-yasg>=1.21  # Swagger
